backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
transformers==4.43.3
torch==2.3.1
mysql-connector-python==9.0.0
gunicorn==22.0.0
